database/raft: save snapshot position first
Save the snapshot position to the WAL before saving the snapshot data to
disk. On boot, we open the WAL at the most recently saved snapshot's
position. If we don't save the position to the WAL first, we might open
the WAL at a snapshot position that was never saved to the WAL.

We might want to wait until the CoreOS folks confirm this in
coreos/etcd#8082. Also, it's debatable whether this meets the
requirements for a point release given that this bug has a very low
probablity of surfacing. WDYT?